fix security issues

+++ mstore-api.php
@@ -78,7 +78,7 @@ class MstoreCheckOut
                 $items = explode("\n", $note);
                 if (strpos($items[0], "URL:") !== false) {
                     $url = str_replace("URL:", "", $items[0]);
-                    echo '<iframe width="600" height="500" src="' . $url . '"></iframe>';
+                    echo '<iframe width="600" height="500" src="' . esc_url($url) . '"></iframe>';
                 }
             }
         }
@@ -127,13 +127,16 @@ class MstoreCheckOut
     }
 
     function mstore_delete_json_file(){
-        $id = $_REQUEST['id'];
+        $id = sanitize_text_field($_REQUEST['id']);
+        $nonce = sanitize_text_field($_REQUEST['nonce']);
         if(strlen($id) == 2){
-            $uploads_dir   = wp_upload_dir();
-            $filePath = trailingslashit( $uploads_dir["basedir"] )."/2000/01/config_".$id.".json";
-            unlink($filePath);
-            echo "success";
-            die();
+            if (wp_verify_nonce($nonce, 'delete_config_json_file')) {
+                $uploads_dir   = wp_upload_dir();
+                $filePath = trailingslashit( $uploads_dir["basedir"] )."/2000/01/config_".$id.".json";
+                unlink($filePath);
+                echo "success";
+                die();
+            }
         }
     }
 
@@ -147,31 +150,31 @@ class MstoreCheckOut
 
     function mstore_update_firebase_server_key()
     {
-        $serverKey = $_REQUEST['serverKey'];
+        $serverKey = sanitize_text_field($_REQUEST['serverKey']);
         update_option("mstore_firebase_server_key", $serverKey);
     }
 
     function mstore_update_new_order_title()
     {
-        $title = $_REQUEST['title'];
+        $title = sanitize_text_field($_REQUEST['title']);
         update_option("mstore_new_order_title", $title);
     }
 
     function mstore_update_new_order_message()
     {
-        $message = $_REQUEST['message'];
+        $message = sanitize_text_field($_REQUEST['message']);
         update_option("mstore_new_order_message", $message);
     }
 
     function mstore_update_status_order_title()
     {
-        $title = $_REQUEST['title'];
+        $title = sanitize_text_field($_REQUEST['title']);
         update_option("mstore_status_order_title", $title);
     }
 
     function mstore_update_status_order_message()
     {
-        $message = $_REQUEST['message'];
+        $message = sanitize_text_field($_REQUEST['message']);
         update_option("mstore_status_order_message", $message);
     }
 
@@ -544,7 +547,7 @@ function prepare_checkout()
             WC()->session->set('chosen_payment_method', $data['payment_method']);
         }
         if (isset($data['customer_note']) && !empty($data['customer_note'])) {
-            $_POST["order_comments"] = $data['customer_note'];
+            $_POST["order_comments"] = sanitize_text_field($data['customer_note']);
             $checkout_fields = WC()->checkout->__get("checkout_fields");
             $checkout_fields["order"] = ["order_comments" => ["type" => "textarea", "class" => [], "label" => "Order notes", "placeholder" => "Notes about your order, e.g. special notes for delivery."]];
             WC()->checkout->__set("checkout_fields", $checkout_fields);
